+++ com.unity.shadergraph/Editor/Data/Graphs/GraphData.cs
@@ -1269,8 +1269,14 @@ namespace UnityEditor.ShaderGraph
         public void OnBeforeSerialize()
         {
             m_SerializableNodes = SerializationHelper.Serialize(GetNodes<AbstractMaterialNode>());
+            var nodes = GetNodes<AbstractMaterialNode>().ToList();
+            nodes.Sort((x1, x2) => x1.guid.CompareTo(x2.guid));
+            m_SerializableNodes = SerializationHelper.Serialize(nodes.AsEnumerable());
+            m_Edges.Sort();
             m_SerializableEdges = SerializationHelper.Serialize<IEdge>(m_Edges);
+            m_Properties.Sort((x1, x2) => x1.guid.CompareTo(x2.guid));
             m_SerializedProperties = SerializationHelper.Serialize<AbstractShaderProperty>(m_Properties);
+            m_Keywords.Sort((x1, x2) => x1.guid.CompareTo(x2.guid));
             m_SerializedKeywords = SerializationHelper.Serialize<ShaderKeyword>(m_Keywords);
             m_ActiveOutputNodeGuidSerialized = m_ActiveOutputNodeGuid == Guid.Empty ? null : m_ActiveOutputNodeGuid.ToString();
         }