+++ com.unity.shadergraph/Editor/Drawing/Views/GraphEditorView.cs
@@ -58,6 +58,8 @@ namespace UnityEditor.ShaderGraph.Drawing
         FloatingWindowsLayout m_FloatingWindowsLayout;
 
         public Action saveRequested { get; set; }
+        
+        public Action saveAsRequested { get; set; }
 
         public Func<bool> isCheckedOut { get; set; }
 
@@ -146,6 +148,11 @@ namespace UnityEditor.ShaderGraph.Drawing
                     {
                         if (saveRequested != null)
                             saveRequested();
+                    }
+                    GUILayout.Space(6);
+                    if (GUILayout.Button("Save As...", EditorStyles.toolbarButton))
+                    {
+                        saveAsRequested();
                     }
                     GUILayout.Space(6);
                     if (GUILayout.Button("Show In Project", EditorStyles.toolbarButton))
@@ -1019,6 +1026,7 @@ namespace UnityEditor.ShaderGraph.Drawing
             if (m_GraphView != null)
             {
                 saveRequested = null;
+                saveAsRequested = null;
                 convertToSubgraphRequested = null;
                 showInProjectRequested = null;
                 isCheckedOut = null;